server: fix ammo getter hash type

+++ server/src/bindings/Player.cpp
@@ -871,7 +871,7 @@ static void AmmoGetter(const v8::FunctionCallbackInfo<v8::Value>& info)
     V8_CHECK_ARGS_LEN(1);
     V8_GET_THIS_BASE_OBJECT(_this, IPlayer);
 
-    uint16_t hash;
+    uint32_t hash;
     if(info[0]->IsNumber())
     {
         V8_ARG_TO_UINT(1, ammoHash);
@@ -892,7 +892,7 @@ static void WeaponAmmoGetter(const v8::FunctionCallbackInfo<v8::Value>& info)
     V8_CHECK_ARGS_LEN(1);
     V8_GET_THIS_BASE_OBJECT(_this, IPlayer);
 
-    uint16_t hash;
+    uint32_t hash;
     if(info[0]->IsNumber())
     {
         V8_ARG_TO_UINT(1, weaponHash);
